simplify and minor refactoring `kafka`
