+++ .gitignore
@@ -1,5 +1,3 @@
-kafka
-zookeeper
 api_data/secret.py
 
 # Byte-compiled / optimized / DLL files
