+++ kafka/consumer.py
@@ -3,11 +3,9 @@ from io import StringIO
 
 import boto3
 import pandas as pd
+
 from kafka import KafkaConsumer
-
-from utils import TOPIC_NAME
-
-BUCKET_NAME = "rtc25"
+from utils.utils import BUCKET_NAME, TOPIC_NAME
 
 s3 = boto3.resource("s3").Bucket(BUCKET_NAME)
 
@@ -16,7 +14,6 @@ bootstrap_servers = ["localhost:9091", "localhost:9092", "localhost:9093"]
 
 
 if __name__ == "__main__":
-
     topicName = TOPIC_NAME
 
     consumer = KafkaConsumer(
@@ -28,12 +25,11 @@ if __name__ == "__main__":
     )
 
     print("Start Consumer")
-    args = {"ACL": "public-read"}
+    args = {"ACL": "public-read"}  # 버켓 설정에 따라 주석처리
 
     for msg in consumer:
         try:
             s3_path = f"{msg.value['key']}_{msg.value['year']}_{msg.value['page']}.csv"
-
             print(f"data get : {s3_path}")
             raw_data = pd.json_normalize(msg.value["data"])
             csv_buffer = StringIO()
@@ -42,4 +38,3 @@ if __name__ == "__main__":
             s3.put_object(Body=csv_buffer.getvalue(), Bucket=BUCKET_NAME, Key=s3_path)
         except Exception as e:
             print(f"error : {e.__str__()}")
-            pass
