+++ kafka/requirements.txt
@@ -1,0 +1,3 @@
+kafka-python
+requests
+pandas
