+++ kafka/utils/secret.py
@@ -1,0 +1,6 @@
+DATA_SEOUL_API_KEY = {}
+
+# 일반 인증키(Encoding)
+DATA_GO_KR_EN = {}
+# 일반 인증키(Decoding )
+DATA_GO_KR_DE = {}
